Rename files to remove pmnist from name
