+++ projects/dendrites/permutedMNIST/experiments/__init__.py
@@ -24,8 +24,8 @@ from .centroid import CONFIGS as CENTROID
 from .batch_mnist import CONFIGS as BATCH_MNIST
 from .sp_context_search import CONFIGS as SP_PROTO
 from .sp_context import CONFIGS as SP_CONTEXT
-from .pmnist_no_dendrites import CONFIGS as PMNIST_NO_DENDRITES
-from .batch_pmnist import CONFIGS as BATCH_PMNIST
+from .no_dendrites import CONFIGS as NO_DENDRITES
+from .batch import CONFIGS as BATCH
 
 """
 Import and collect all experiment configurations into one CONFIG
@@ -39,5 +39,5 @@ CONFIGS.update(CENTROID)
 CONFIGS.update(BATCH_MNIST)
 CONFIGS.update(SP_PROTO)
 CONFIGS.update(SP_CONTEXT)
-CONFIGS.update(PMNIST_NO_DENDRITES)
-CONFIGS.update(BATCH_PMNIST)
+CONFIGS.update(NO_DENDRITES)
+CONFIGS.update(BATCH)
