+++ projects/dendrites/permutedMNIST/experiments/batch.py
