+++ projects/dendrites/permutedMNIST/experiments/no_dendrites.py
